Refactor TermInnerRow component and add descriptor_row class
This commit refactors the TermInnerRow component in the src/components/TermInnerRow.jsx file. It adds the descriptor_row class to the td element, and updates the styling in the StyleSheet.css file by adding the .descriptor_row selector. This change adds a left padding of 25px to the descriptor_row class.